A data.frame of geographical zones taken from bbbike.org

Start bicycle routing for... everywhere!

bbbike_zones

Format

A sf dataframe with 430 rows and 14 columns:

name

The, usually English, long-form name of the city

last_modified

When was it last modified?

type

empty

pbf_size_size

Size of the pbf file in bytes.

base_url

The base URL for the city

poly_url

The .poly file location

pbf

Link to the latest .osm.pbf file for this region.

level

An integer code which is always equal to 3L since the bbbike data represent non-hierarchical geographical zones. This is used only for matching operations in case of spatial input. The oe_* functions will select the geographical area closest to the input place with the highest "level". See ?geofabrik_zones for an example of hierarchical structure.

geometry

The sfc for that geographical region, rectangular.

Source

https://download.bbbike.org/osm/

Details

A sf data.frame containing the URLs, names and file_size of the OSM extracts. See https://download.bbbike.org/osm/.