An sf object of geographical zones taken from bbbike.org

Start bicycle routing for... everywhere!

bbbike_zones

Format

An sf object with 235 rows and 11 columns:

name

The, usually English, long-form name of the city.

last_modified

When was it last modified?

type

empty

pbf_file_size

Size of the pbf file in bytes.

base_url

The base URL for the city.

poly_url

The .poly file location.

pbf

Link to the latest .osm.pbf file for this region.

level

An integer code always equal to 3 (since the bbbike data represent non-hierarchical geographical zones). This is used only for matching operations in case of spatial input. The oe_* functions will select the geographical area closest to the input place with the highest "level". See geofabrik_zones for an example of a hierarchical structure.

geometry

The sfg for that geographical region, rectangular.

Source

https://download.bbbike.org/osm/

Details

An sf object containing the URLs, names and file_size of the OSM extracts.

See also