Get the location of files

This function takes a place name and it returns the path of .pbf and .gpkg files associated with it.

oe_find(
  place,
  provider = "geofabrik",
  download_directory = oe_download_directory(),
  download_if_missing = FALSE,
  quiet = FALSE,
  ...
)

Arguments

place
Description of the geographical area that should be matched with a .osm.pbf file through the chosen provider. Can be either a length-1 character vector, an sf or sfc object, or a numeric vector of coordinates with length 2. In the last case, it is assumed that the EPSG code is 4326 specified as c(LON, LAT), while you can use any CRS with an sf or sfc object. See Details and examples in oe_match().

provider
Which provider should be used to download the data? Available providers can be found with the following command: oe_providers(). For oe_get() and oe_match(), if place is equal to ITS Leeds, then provider is set equal to test. This is just for simple examples and internal tests.

download_directory
Directory where the files downloaded by osmextract are stored. By default it is equal to oe_download_directory().

download_if_missing
Attempt to download the file if it cannot be found? FALSE by default.

quiet
Boolean. If FALSE, the function prints informative messages. Starting from sf version 0.9.6, if quiet is equal to FALSE, then vectortranslate operations will display a progress bar.

...
Extra arguments that are passed to oe_match() and oe_get(). Please note that you cannot modify the argument download_only.

Value

A character vector of length one (or two) representing the path(s) of the corresponding .pbf (and .gpkg) files.

Details

The matching between the existing files (saved in a directory specified by download_directory parameter) and the input place is performed using list.files, setting a pattern equal to the basename of the URL associated to the input place. For example, if you specify place = "Isle of Wight", then the input place is matched with a URL of a .osm.pbf file (via oe_match()) and the matching is performed setting a pattern equal to the basename of that URL.

If there is no file in download_directory that can be matched with the basename and download_if_missing parameter is equal to TRUE, then the function tries to download and translate a new file from the chosen provider (geofabrik is the default provider). If download_if_missing parameter is equal to FALSE (default value), then the function stops with an error.

Examples

res = oe_get("ITS Leeds", quiet = TRUE, download_directory = tempdir())
oe_find("ITS Leeds", provider = "test", download_directory = tempdir())
#> The input place was matched with: ITS Leeds
#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmptxGtQu/test_its-example.gpkg"   
#> [2] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmptxGtQu/test_its-example.osm.pbf"

if (FALSE) {
  oe_find("Isle of Wight", download_directory = tempdir())
  oe_find("Malta", download_if_missing = TRUE, download_directory = tempdir())
  oe_find(
    "Leeds",
    provider = "bbbike",
    download_if_missing = TRUE,
    download_directory = tempdir()
  )}
# Remove .pbf and .gpkg files in tempdir
# (since they may interact with other examples)
file.remove(list.files(path = tempdir(), pattern = "(pbf|gpkg)", full.names = TRUE))
#> [1] TRUE TRUE