Download, translate and read OSM extracts from several providers

This function is used to download, translate and read OSM extracts obtained from several providers. It is a wrapper around oe_match() and oe_read(). Check the introductory vignette, the examples and the help pages of the wrapped functions to understand the details behind all parameters.

oe_get(
  place,
  layer = "lines",
  ...,
  provider = "geofabrik",
  match_by = "name",
  max_string_dist = 1,
  interactive_ask = FALSE,
  download_directory = oe_download_directory(),
  force_download = FALSE,
  max_file_size = 5e+08,
  vectortranslate_options = NULL,
  osmconf_ini = NULL,
  extra_tags = NULL,
  force_vectortranslate = FALSE,
  download_only = FALSE,
  skip_vectortranslate = FALSE,
  quiet = TRUE
)

Arguments

place
Description of the geographical area that should be matched with a .osm.pbf file through the chosen provider. Can be either a length-1 character vector, a length-1 sfc_POINT object or a numeric vector of coordinates with length 2. In the latter case it is assumed that the EPSG code is 4326, while you can use any EPSG code with an sfc_POINT object. See Details and examples in oe_match().

layer
Which layer should be read in? Typically points, lines (the default), multilinestrings, multipolygons or other_relations.

...
Arguments that will be passed to sf::st_read(), like query or stringsAsFactors. Check the introductory vignette to understand how to create your own (SQL-like) queries.

provider
Which provider should be used to download the data? Available providers can be found with the following command: oe_providers().

match_by
Which column of the provider's database should be used for matching the input place with a .osm.pbf file? The default is "name". Check details and examples in oe_match() to understand how this parameter works. Ignored if place is not a character vector since the matching is performed through a spatial operation.

max_string_dist
Numerical value greater or equal than 0. What is the maximum distance in fuzzy matching (i.e. Approximate String Distance, see adist()) between input place and match_by column to tolerate before asking the user to select which zone to download? This parameter is set equal to 0 if match_by is equal to iso3166_1_alpha2 or iso3166_2. Check Details and examples in oe_match() to understand why this parameter is important. Ignored if place is not a character vector since the matching is performed through a spatial operation.

interactive_ask
Boolean. If TRUE the function creates an interactive menu in case the best match is further than max_string_dist, otherwise it fails with stop(). Check details and examples in oe_match() to understand why this parameter is important. Ignored if place is not a character vector since the matching is performed through a spatial operation.

download_directory
Where to download the file containing the OSM data? By default this is equal to oe_download_directory(), which is equal to tempdir() and it changes each time you restart R. You can set a persistent download_directory by adding the following to your .Renviron file (e.g. with usethis::edit_r_environ()): OSMEXT_DOWNLOAD_DIRECTORY=/path/to/osm/data.

force_download
Should the .osm.pbf file be updated if it has already been downloaded? FALSE by default. This parameter is used to update old .osm.pbf files.

max_file_size
The maximum file size to download without asking in interactive mode. Default: 5e+8, half a gigabyte.

vectortranslate_options
Options to pass to the sf::gdal_utils() argument options. Set by default. Check Details in the introductory vignette and in oe_vectortranslate().

osmconf_ini
The configuration file specifying which columns should be in the resulting data frame. See documentation at gdal.org. Check Details in oe_vectortranslate().

extra_tags
Which addition columns, corresponding to OSM tags, should be in the resulting dataset? FALSE by default. Check Details at oe_vectortranslate() and oe_get_keys() .

force_vectortranslate
Boolean. Force the original .pbf file to be translated into a .gpkg file, even if a .gpkg with the same name already exists? Check Details at oe_vectortranslate() .

download_only
Boolean. If TRUE then the function only returns the path where the matched file is stored, instead of reading it. FALSE by default.

skip_vectortranslate
Boolean. If TRUE then the function skips all vectortranslate operations and it reads (or simply returns the path) of the .osm.pbf file. FALSE by default.

quiet
Boolean. If FALSE the function prints informative messages.

Value

An sf object.

Details

The algorithm that we use for importing an OSM extract data into R is divided into 4 steps: 1) Matching the input place with the url of a .pbf file; 2) download the .pbf file; 3) convert it into .gpkg format and 4) read-in the .gpkg file. The function oe_match() is used to perform the first operation and the function oe_read() (which is a wrapper around oe_download(), oe_vectortranslate() and sf::st_read()) performs the other three operations.

See also

Examples

# Download OSM extracts associated to a simple test.
its = oe_get("ITS Leeds", provider = "test", quiet = FALSE)

#> The input place was matched with: ITS Leeds

#> The chosen file was already detected in the download directory. Skip downloading.

#> Start with the vectortranslate operations on the input file!

#> Finished the vectortranslate operations on the input file!

#> Reading layer `lines' from data source `/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpjl9VqC/test_its-example.gpkg' using driver `GPKG'
#> Simple feature collection with 189 features and 9 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.562458 ymin: 53.80471 xmax: -1.548076 ymax: 53.81105
#> geographic CRS: WGS 84

class(its)

#> [1] "sf"         "data.frame"

summary(sf::st_geometry_type(its))

#>           GEOMETRY              POINT         LINESTRING            POLYGON 
#>                  0                  0                189                  0 
#>         MULTIPOINT    MULTILINESTRING       MULTIPOLYGON GEOMETRYCOLLECTION 
#>                  0                  0                  0                  0 
#>     CIRCULARSTRING      COMPOUNDCURVE       CURVEPOLYGON         MULTICURVE 
#>                  0                  0                  0                  0 
#>       MULTISURFACE              CURVE            SURFACE  POLYHEDRALSURFACE 
#>                  0                  0                  0                  0 
#>                TIN           TRIANGLE 
#>                  0                  0 

# Add another layer to the test file
its_points = oe_get(
"ITS Leeds",
provider = "test",
layer = "points",
quiet = FALSE
)

#> The input place was matched with: ITS Leeds

#> The chosen file was already detected in the download directory. Skip downloading.

#> Adding a new layer to the .gpkg file

#> Start with the vectortranslate operations on the input file!

#> Finished the vectortranslate operations on the input file!

#> Reading layer `points' from data source `/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpjl9VqC/test_its-example.gpkg' using driver `GPKG'
#> Simple feature collection with 186 features and 10 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -1.568766 ymin: 53.80569 xmax: -1.549451 ymax: 53.81136
#> geographic CRS: WGS 84

summary(sf::st_geometry_type(its_points))

#>           GEOMETRY              POINT         LINESTRING            POLYGON 
#>                  0                186                  0                  0 
#>         MULTIPOINT    MULTILINESTRING       MULTIPOLYGON GEOMETRYCOLLECTION 
#>                  0                  0                  0                  0 
#>     CIRCULARSTRING      COMPOUNDCURVE       CURVEPOLYGON         MULTICURVE 
#>                  0                  0                  0                  0 
#>       MULTISURFACE              CURVE            SURFACE  POLYHEDRALSURFACE 
#>                  0                  0                  0                  0 
#>                TIN           TRIANGLE 
#>                  0                  0 

# Get the .osm.pbf and .gpkg file path
oe_get("ITS Leeds", provider = "test", download_only = TRUE)

#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmpjl9VqC/test_its-example.gpkg"

oe_get(
"ITS Leeds",
provider = "test",
download_only = TRUE,
skip_vectortranslate = TRUE
)

#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmpjl9VqC/test_its-example.osm.pbf"

# Add additional tags
im = oe_get(
"ITS Leeds",
provider = "test",
extra_tags = "oneway",
quiet = FALSE
)

#> The input place was matched with: ITS Leeds

#> The chosen file was already detected in the download directory. Skip downloading.

#> Start with the vectortranslate operations on the input file!

#> Finished the vectortranslate operations on the input file!

#> Reading layer `lines' from data source `/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpjl9VqC/test_its-example.gpkg' using driver `GPKG'
#> Simple feature collection with 189 features and 10 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.562458 ymin: 53.80471 xmax: -1.548076 ymax: 53.81105
#> geographic CRS: WGS 84

names(im)

#>  [1] "osm_id"     "name"       "highway"    "waterway"   "aerialway" 
#>  [6] "barrier"    "man_made"   "oneway"     "z_order"    "other_tags"
#> [11] "geometry"  

# The ITS Leeds file is stored on the github page of the package, so we
# need to manually set the provider. This is not the case with other
# standards providers.

if (FALSE) {
west_yorkshire = oe_get("West Yorkshire", quiet = FALSE)
# If you run it again, the function will not download the file or convert it
west_yorkshire = oe_get("West Yorkshire", quiet = FALSE)
# match with coordinates (any EPSG)
milan_duomo = sf::st_sfc(sf::st_point(c(1514924, 5034552)), crs = 3003)
# Warning: the .pbf file is 400MB
oe_get(milan_duomo, quiet = FALSE)
# match with numeric coordinates (EPSG = 4326)
oe_match(c(9.1916, 45.4650), quiet = FALSE)
# alternative providers
baku = oe_get(place = "Baku", provider = "bbbike", quiet = FALSE)}