Find, download, translate and read OSM extracts from several providers

This function is used to find, download, translate and read OSM extracts obtained from several providers. It is a wrapper around oe_match() and oe_read(). Check the introductory vignette, the examples and the help pages of the wrapped functions to understand the details behind all parameters.

oe_get(
  place,
  layer = "lines",
  ...,
  provider = "geofabrik",
  match_by = "name",
  max_string_dist = 1,
  level = NULL,
  download_directory = oe_download_directory(),
  force_download = FALSE,
  max_file_size = 5e+08,
  vectortranslate_options = NULL,
  osmconf_ini = NULL,
  extra_tags = NULL,
  force_vectortranslate = FALSE,
  download_only = FALSE,
  skip_vectortranslate = FALSE,
  never_skip_vectortranslate = FALSE,
  quiet = FALSE
)

Arguments

place
Description of the geographical area that should be matched with a .osm.pbf file through the chosen provider. Can be either a length-1 character vector, an sf or sfc object, or a numeric vector of coordinates with length 2. In the last case, it is assumed that the EPSG code is 4326 specified as c(LON, LAT), while you can use any CRS with an sf or sfc object. See Details and examples in oe_match().

layer
Which layer should be read in? Typically points, lines (the default), multilinestrings, multipolygons or other_relations. If you specify an ad-hoc query using the argument query (see introductory vignette and examples), then oe_get() and oe_read() will read the layer specified in the query and ignore layer. See also #122.

...
Arguments that will be passed to sf::st_read(), like query, wkt_filter or stringsAsFactors. Check the introductory vignette to understand how to create your own (SQL-like) queries.

provider
Which provider should be used to download the data? Available providers can be found with the following command: oe_providers(). For oe_get() and oe_match(), if place is equal to ITS Leeds, then provider is set equal to test. This is just for simple examples and internal testings.

match_by
Which column of the provider's database should be used for matching the input place with a .osm.pbf file? The default is "name". Check Details and Examples in oe_match() to understand how this parameter works. Ignored if place is not a character vector since the matching is performed through a spatial operation.

max_string_dist
Numerical value greater or equal than 0. What is the maximum distance in fuzzy matching (i.e. Approximate String Distance, see adist()) between input place and match_by column to tolerate before testing alternative providers or looking for geographical matching with Nominatim API? This parameter is set equal to 0 if match_by is equal to iso3166_1_alpha2 or iso3166_2. Check Details and Examples in oe_match() to understand why this parameter is important. Ignored if place is not a character vector since the matching is performed through a spatial operation.

level
An integer representing the desired hierarchical level in case of spatial matching. For the geofabrik provider, for example, 1 corresponds with continent-level datasets, 2 for countries, 3 corresponds to regions and 4 to subregions. Hence, we could approximately say that smaller administrative units correspond to bigger levels. If NULL, the default, the oe_* functions will select the highest available level. See Details and Examples in oe_match().

download_directory
Where to download the file containing the OSM data? By default this is equal to oe_download_directory(), which is equal to tempdir() and it changes each time you restart R. You can set a persistent download_directory by adding the following to your .Renviron file (e.g. with edit_r_environ function in usethis package): OSMEXT_DOWNLOAD_DIRECTORY=/path/to/osm/data.

force_download
Should the .osm.pbf file be updated if it has already been downloaded? FALSE by default. This parameter is used to update old .osm.pbf files.

max_file_size
The maximum file size to download without asking in interactive mode. Default: 5e+8, half a gigabyte.

vectortranslate_options
Options passed to the sf::gdal_utils() argument options. Set by default. Check details in the introductory vignette and the help page of oe_vectortranslate().

osmconf_ini
The configuration file. See documentation at gdal.org. Check details in the introductory vignette and the help page of oe_vectortranslate(). Set by default.

extra_tags
Which additional columns, corresponding to OSM tags, should be in the resulting dataset? NULL by default. Check the introductory vignette and the help pages of oe_vectortranslate() and oe_get_keys(). Ignored when osmconf_ini is not NULL.

force_vectortranslate
Boolean. Force the original .pbf file to be translated into a .gpkg file, even if a .gpkg with the same name already exists? Check the introductory vignette and the help page of oe_vectortranslate().

download_only
Boolean. If TRUE, then the function only returns the path where the matched file is stored, instead of reading it. FALSE by default.

skip_vectortranslate
Boolean. If TRUE, then the function skips all vectortranslate operations and it reads (or simply returns the path) of the .osm.pbf file. FALSE by default.

never_skip_vectortranslate
Boolean. This is used in case the user passed its own .ini file or vectortranslate options (since, in those case, it's too difficult to determine if an existing .gpkg file was generated following the same options.)

quiet
Boolean. If FALSE, the function prints informative messages. Starting from sf version 0.9.6, if quiet is equal to FALSE, then vectortranslate operations will display a progress bar.

Value

An sf object.

Details

The algorithm that we use for importing an OSM extract data into R is divided into 4 steps: 1) match the input place with the url of a .pbf file; 2) download the .pbf file; 3) convert it into .gpkg format and 4) read-in the .gpkg file. The function oe_match() is used to perform the first operation and the function oe_read() (which is a wrapper around oe_download(), oe_vectortranslate() and sf::st_read()) performs the other three operations.

See also

Examples

# Download OSM extracts associated to a simple test.
its = oe_get("ITS Leeds", quiet = FALSE)

#> The input place was matched with: ITS Leeds

#> The chosen file was already detected in the download directory. Skip downloading.

#> The corresponding gpkg file was already detected. Skip vectortranslate operations.

#> Reading layer `lines' from data source `/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpXQfWXu/test_its-example.gpkg' using driver `GPKG'
#> Simple feature collection with 189 features and 9 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.562458 ymin: 53.80471 xmax: -1.548076 ymax: 53.81105
#> geographic CRS: WGS 84

class(its)

#> [1] "sf"         "data.frame"

unique(sf::st_geometry_type(its))

#> [1] LINESTRING
#> 18 Levels: GEOMETRY POINT LINESTRING POLYGON MULTIPOINT ... TRIANGLE

# Get another layer from the test extract
its_points = oe_get("ITS Leeds", layer = "points")

#> The input place was matched with: ITS Leeds

#> The chosen file was already detected in the download directory. Skip downloading.

#> Adding a new layer to the .gpkg file

#> Start with the vectortranslate operations on the input file!

#> 0...10...20...30...40...50...60...70...80...90...100 - done.

#> Finished the vectortranslate operations on the input file!

#> Reading layer `points' from data source `/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpXQfWXu/test_its-example.gpkg' using driver `GPKG'
#> Simple feature collection with 186 features and 10 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -1.568766 ymin: 53.80569 xmax: -1.549451 ymax: 53.81136
#> geographic CRS: WGS 84

unique(sf::st_geometry_type(its_points))

#> [1] POINT
#> 18 Levels: GEOMETRY POINT LINESTRING POLYGON MULTIPOINT ... TRIANGLE

# Get the .osm.pbf and .gpkg file path
oe_get("ITS Leeds", download_only = TRUE)

#> The input place was matched with: ITS Leeds

#> The chosen file was already detected in the download directory. Skip downloading.

#> The corresponding gpkg file was already detected. Skip vectortranslate operations.

#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpXQfWXu/test_its-example.gpkg"

oe_get("ITS Leeds", download_only = TRUE, skip_vectortranslate = TRUE)

#> The input place was matched with: ITS Leeds

#> The chosen file was already detected in the download directory. Skip downloading.

#> [1] "/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpXQfWXu/test_its-example.osm.pbf"

# See also ?oe_find()

# Add additional tags
its_with_oneway = oe_get("ITS Leeds", extra_tags = "oneway", quiet = FALSE)

#> The input place was matched with: ITS Leeds

#> The chosen file was already detected in the download directory. Skip downloading.

#> Start with the vectortranslate operations on the input file!

#> 0...10...20...30...40...50...60...70...80...90...100 - done.

#> Finished the vectortranslate operations on the input file!

#> Reading layer `lines' from data source `/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpXQfWXu/test_its-example.gpkg' using driver `GPKG'
#> Simple feature collection with 189 features and 10 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.562458 ymin: 53.80471 xmax: -1.548076 ymax: 53.81105
#> geographic CRS: WGS 84

names(its_with_oneway)

#>  [1] "osm_id"     "name"       "highway"    "waterway"   "aerialway" 
#>  [6] "barrier"    "man_made"   "oneway"     "z_order"    "other_tags"
#> [11] "geometry"  

table(its_with_oneway$oneway, useNA = "ifany")

#> 
#>  yes <NA> 
#>   13  176 

# Use the query argument to get only oneway streets:
q = "SELECT * FROM 'lines' WHERE oneway IN ('yes')"
its_oneway = oe_get("ITS Leeds", query = q)

#> The input place was matched with: ITS Leeds

#> The chosen file was already detected in the download directory. Skip downloading.

#> The corresponding gpkg file was already detected. Skip vectortranslate operations.

#> Reading layer `lines' from data source `/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpXQfWXu/test_its-example.gpkg' using driver `GPKG'
#> Simple feature collection with 13 features and 10 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.558828 ymin: 53.80595 xmax: -1.548899 ymax: 53.809
#> geographic CRS: WGS 84

its_oneway

#> Simple feature collection with 13 features and 10 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.558828 ymin: 53.80595 xmax: -1.548899 ymax: 53.809
#> geographic CRS: WGS 84
#> First 10 features:
#>       osm_id                    name     highway waterway aerialway barrier
#> 1    6277600          Cavendish Road     service     <NA>      <NA>    <NA>
#> 2    6277601          Cavendish Road residential     <NA>      <NA>    <NA>
#> 3    6295680           Blenheim Walk       trunk     <NA>      <NA>    <NA>
#> 4   31705836          Woodhouse Lane       trunk     <NA>      <NA>    <NA>
#> 5   31705838                    <NA>  trunk_link     <NA>      <NA>    <NA>
#> 6   38422788           Blenheim Walk       trunk     <NA>      <NA>    <NA>
#> 7  147151516 Back Westbourne Terrace     service     <NA>      <NA>    <NA>
#> 8  151645336                    <NA> residential     <NA>      <NA>    <NA>
#> 9  160502811                    <NA>     service     <NA>      <NA>    <NA>
#> 10 552695946                    <NA>  trunk_link     <NA>      <NA>    <NA>
#>    man_made oneway z_order
#> 1      <NA>    yes       0
#> 2      <NA>    yes       3
#> 3      <NA>    yes       8
#> 4      <NA>    yes       8
#> 5      <NA>    yes       8
#> 6      <NA>    yes       8
#> 7      <NA>    yes       0
#> 8      <NA>    yes       3
#> 9      <NA>    yes       0
#> 10     <NA>    yes       8
#>                                                                                                        other_tags
#> 1                                                                             "access"=>"permissive","lanes"=>"1"
#> 2                                                                                                    "lanes"=>"1"
#> 3  "lanes"=>"2","lit"=>"yes","maxspeed"=>"30 mph","ref"=>"A660","turn:lanes"=>"left;through|through;slight_right"
#> 4      "lanes"=>"2","lanes:psv"=>"1","lit"=>"yes","maxspeed"=>"30 mph","ref"=>"A660","turn:lanes"=>"none|through"
#> 5                                                                               "lanes"=>"1","maxspeed"=>"30 mph"
#> 6                                                "bicycle"=>"yes","lanes"=>"2","maxspeed"=>"30 mph","ref"=>"A660"
#> 7                    "motor_vehicle"=>"destination","source:name"=>"OS_OpenData_Locator","surface"=>"cobblestone"
#> 8                                                                                                    "lanes"=>"1"
#> 9                                                                                                            <NA>
#> 10                                                                                                   "lanes"=>"1"
#>                          geometry
#> 1  LINESTRING (-1.552587 53.80...
#> 2  LINESTRING (-1.551771 53.80...
#> 3  LINESTRING (-1.552894 53.80...
#> 4  LINESTRING (-1.552349 53.80...
#> 5  LINESTRING (-1.552261 53.80...
#> 6  LINESTRING (-1.552261 53.80...
#> 7  LINESTRING (-1.557987 53.80...
#> 8  LINESTRING (-1.552181 53.80...
#> 9  LINESTRING (-1.553221 53.80...
#> 10 LINESTRING (-1.551952 53.80...

if (FALSE) {
# A more complex example
west_yorkshire = oe_get("West Yorkshire", quiet = FALSE)
# If you run it again, the function will not download the file
# or convert it again
west_yorkshire = oe_get("West Yorkshire", quiet = FALSE)
# Match with place name
oe_get("Milan") # Warning: the .pbf file is 400MB
oe_get("Vatican City")
oe_get("Zurich")

# Match with coordinates (any EPSG)
milan_duomo = sf::st_sfc(sf::st_point(c(1514924, 5034552)), crs = 3003)
oe_get(milan_duomo, quiet = FALSE) # Warning: the .pbf file is 400MB
# Match with numeric coordinates (EPSG = 4326)
oe_match(c(9.1916, 45.4650), quiet = FALSE)
# Alternative providers
baku = oe_get(place = "Baku", provider = "bbbike", quiet = FALSE)

# Other examples:
oe_get("RU", match_by = "iso3166_1_alpha2", quiet = FALSE)
# The following example mimics read_sf
oe_get("Andora", stringsAsFactors = FALSE, quiet = TRUE, as_tibble = TRUE)}