Return all keys stored in "other_tags" column

This function is used to return the names of all keys that are stored in "other_tags" column since they were not explicitly included in the file. See Details.

oe_get_keys(zone, layer = "lines")

# S3 method for default
oe_get_keys(zone, layer = "lines")

# S3 method for character
oe_get_keys(zone, layer = "lines")

# S3 method for sf
oe_get_keys(zone, layer = "lines")

Arguments

zone
An sf object (typically read-in with oe_read() or oe_get()) with an other_tags field, or a character vector (of length 1) that points to a .gpkg file (typically created using oe_vectortranslate() or oe_get()).

layer
Which layer should be read in? Typically points, lines (the default), multilinestrings, multipolygons or other_relations.

Value

A character vector indicating the name of all keys stored in "other_tags" field.

Details

OSM data are typically documented using several tags. A tag is a pair of two items, namely a key and a value. As we documented in oe_vectortranslate(), the conversion between .osm.pbf and .gpkg formats is governed by a CONFIG file that indicates which tags are explicitly added to the .gpkg file. All the other keys stored in the .osm.pbf file are automatically appended using an "other_tags" field, with a syntax compatible with the PostgreSQL HSTORE type. This function is used to display the names of all keys stored in the "other_tags" column.

You can also use the hstore_get_value() function from GDAL to extract one particular tag from an existing .gpkg file. Check the introductory vignette and see examples.

The definition of a generic S3 implementation started in osmextract/issues/138.

See also

Examples

itsleeds_gpkg_path = oe_get("ITS Leeds", download_only = TRUE)

#> The input place was matched with: ITS Leeds

#> The chosen file was already detected in the download directory. Skip downloading.

#> The corresponding gpkg file was already detected. Skip vectortranslate operations.

oe_get_keys(itsleeds_gpkg_path)

#>  [1] "bicycle"             "foot"                "website"            
#>  [4] "access"              "lanes"               "lit"                
#>  [7] "maxspeed"            "ref"                 "turn:lanes"         
#> [10] "service"             "surface"             "motor_vehicle"      
#> [13] "source:name"         "lanes:psv"           "lanes:backward"     
#> [16] "lanes:forward"       "lanes:psv:backward"  "turn:lanes:forward" 
#> [19] "covered"             "layer"               "lcn"                
#> [22] "natural"             "step_count"          "frequency"          
#> [25] "operator"            "power"               "source:geometry"    
#> [28] "substation"          "voltage"             "tunnel"             
#> [31] "indoor"              "level"               "maxheight"          
#> [34] "incline"             "bridge"              "alt_name"           
#> [37] "turn:lanes:backward"

itsleeds = oe_get("ITS Leeds")

#> The input place was matched with: ITS Leeds

#> The chosen file was already detected in the download directory. Skip downloading.

#> The corresponding gpkg file was already detected. Skip vectortranslate operations.

#> Reading layer `lines' from data source `/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpCvPc3g/test_its-example.gpkg' using driver `GPKG'
#> Simple feature collection with 189 features and 10 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.562458 ymin: 53.80471 xmax: -1.548076 ymax: 53.81105
#> geographic CRS: WGS 84

oe_get_keys(itsleeds)

#>  [1] "bicycle"             "foot"                "website"            
#>  [4] "access"              "lanes"               "lit"                
#>  [7] "maxspeed"            "ref"                 "turn:lanes"         
#> [10] "service"             "surface"             "motor_vehicle"      
#> [13] "source:name"         "lanes:psv"           "lanes:backward"     
#> [16] "lanes:forward"       "lanes:psv:backward"  "turn:lanes:forward" 
#> [19] "covered"             "layer"               "lcn"                
#> [22] "natural"             "step_count"          "frequency"          
#> [25] "operator"            "power"               "source:geometry"    
#> [28] "substation"          "voltage"             "tunnel"             
#> [31] "indoor"              "level"               "maxheight"          
#> [34] "incline"             "bridge"              "alt_name"           
#> [37] "turn:lanes:backward"

# Add an extra key to an existing .gpkg file without vectortranslate
names(oe_read(
  itsleeds_gpkg_path,
  query = "SELECT *,  hstore_get_value(other_tags, 'oneway')  AS oneway FROM lines"
))

#> Reading layer `lines' from data source `/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpCvPc3g/test_its-example.gpkg' using driver `GPKG'
#> Simple feature collection with 189 features and 10 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.562458 ymin: 53.80471 xmax: -1.548076 ymax: 53.81105
#> geographic CRS: WGS 84

#>  [1] "osm_id"     "name"       "highway"    "waterway"   "aerialway" 
#>  [6] "barrier"    "man_made"   "oneway"     "z_order"    "other_tags"
#> [11] "geometry"