Return all keys stored in "other_tags" column

This function is used to return the names of all keys that are stored in "other_tags" column since they were not explicitly included in the file. See Details.

oe_get_keys(zone, layer = "lines")

# S3 method for default
oe_get_keys(zone, layer = "lines")

# S3 method for character
oe_get_keys(zone, layer = "lines")

# S3 method for sf
oe_get_keys(zone, layer = "lines")

Arguments

zone
An sf object (typically read-in with oe_read() or oe_get()) with an other_tags field, or a character vector (of length 1) that points to a .gpkg file (typically created using oe_vectortranslate() or oe_get()).

layer
Which layer should be read in? Typically points, lines (the default), multilinestrings, multipolygons or other_relations. If you specify an ad-hoc query using the argument query (see introductory vignette and examples), then oe_get() and oe_read() will read the layer specified in the query and ignore layer. See also #122.

Value

A character vector indicating the name of all keys stored in "other_tags" field.

Details

OSM data are typically documented using several tags. A tag is a pair of two items, namely a key and a value. As we documented in oe_vectortranslate(), the conversion between .osm.pbf and .gpkg formats is governed by a CONFIG file that indicates which tags are explicitly added to the .gpkg file. All the other keys stored in the .osm.pbf file are automatically appended using an "other_tags" field, with a syntax compatible with the PostgreSQL HSTORE type. This function is used to display the names of all keys stored in the "other_tags" column.

You can also use the hstore_get_value() function from GDAL to extract one particular tag from an existing .gpkg file. Check the introductory vignette and see examples.

The definition of a generic S3 implementation started in osmextract/issues/138.

See also

Examples

itsleeds_gpkg_path = oe_get(
  "ITS Leeds",
  download_only = TRUE,
  download_directory = tempdir(),
  quiet = TRUE
)
oe_get_keys(itsleeds_gpkg_path)

#>  [1] "bicycle"             "foot"                "website"            
#>  [4] "access"              "lanes"               "oneway"             
#>  [7] "lit"                 "maxspeed"            "ref"                
#> [10] "turn:lanes"          "service"             "surface"            
#> [13] "motor_vehicle"       "source:name"         "lanes:psv"          
#> [16] "lanes:backward"      "lanes:forward"       "lanes:psv:backward" 
#> [19] "turn:lanes:forward"  "covered"             "layer"              
#> [22] "lcn"                 "natural"             "step_count"         
#> [25] "frequency"           "operator"            "power"              
#> [28] "source:geometry"     "substation"          "voltage"            
#> [31] "tunnel"              "indoor"              "level"              
#> [34] "maxheight"           "incline"             "bridge"             
#> [37] "alt_name"            "turn:lanes:backward"

itsleeds = oe_get("ITS Leeds", quiet = TRUE, download_directory = tempdir())
oe_get_keys(itsleeds)

#>  [1] "bicycle"             "foot"                "website"            
#>  [4] "access"              "lanes"               "oneway"             
#>  [7] "lit"                 "maxspeed"            "ref"                
#> [10] "turn:lanes"          "service"             "surface"            
#> [13] "motor_vehicle"       "source:name"         "lanes:psv"          
#> [16] "lanes:backward"      "lanes:forward"       "lanes:psv:backward" 
#> [19] "turn:lanes:forward"  "covered"             "layer"              
#> [22] "lcn"                 "natural"             "step_count"         
#> [25] "frequency"           "operator"            "power"              
#> [28] "source:geometry"     "substation"          "voltage"            
#> [31] "tunnel"              "indoor"              "level"              
#> [34] "maxheight"           "incline"             "bridge"             
#> [37] "alt_name"            "turn:lanes:backward"

# Add an extra key to an existing .gpkg file without vectortranslate
names(oe_read(
  itsleeds_gpkg_path,
  query = "SELECT *,  hstore_get_value(other_tags, 'oneway')  AS oneway FROM lines"
))

#> Reading layer `lines' from data source `/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpy6I0Z6/test_its-example.gpkg' using driver `GPKG'
#> Simple feature collection with 189 features and 10 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -1.562458 ymin: 53.80471 xmax: -1.548076 ymax: 53.81105
#> Geodetic CRS:  WGS 84

#>  [1] "osm_id"     "name"       "highway"    "waterway"   "aerialway" 
#>  [6] "barrier"    "man_made"   "z_order"    "other_tags" "oneway"    
#> [11] "geometry"  

# Remove .pbf and .gpkg files in tempdir
# (since they may interact with other examples)
file.remove(list.files(path = tempdir(), pattern = "(pbf|gpkg)", full.names = TRUE))

#> [1] TRUE TRUE