Match input place with a URL

This function is used to match an input place with the URL of a .osm.pbf file (and its file-size, if present). The URLs are stored in several provider's databases. See oe_providers() and examples.

oe_match(place, ...)

# S3 method for default
oe_match(place, ...)

# S3 method for sfc_POINT
oe_match(place, provider = "geofabrik", quiet = FALSE, ...)

# S3 method for numeric
oe_match(place, provider = "geofabrik", quiet = FALSE, ...)

# S3 method for character
oe_match(
  place,
  provider = "geofabrik",
  quiet = FALSE,
  match_by = "name",
  max_string_dist = 1,
  ...
)

Arguments

place
Description of the geographical area that should be matched with a .osm.pbf file through the chosen provider. Can be either a length-1 character vector, a length-1 sfc_POINT object or a numeric vector of coordinates with length 2. In the latter case it is assumed that the EPSG code is 4326 specified as c(LON, LAT), while you can use any CRS with an sfc_POINT object. See Details and examples in oe_match().

...
arguments passed to other methods

provider
Which provider should be used to download the data? Available providers can be found with the following command: oe_providers(). For oe_get() and oe_match(), if place is equal to ITS Leeds, then provider is set equal to test. This is just for simple examples and internal testings.

quiet
Boolean. If FALSE the function prints informative messages. Starting from sf version 0.9.6, if quiet is equal to FALSE, then vectortranslate operations will display a progress bar.

match_by
Which column of the provider's database should be used for matching the input place with a .osm.pbf file? The default is "name". Check Details and Examples in oe_match() to understand how this parameter works. Ignored if place is not a character vector since the matching is performed through a spatial operation.

max_string_dist
Numerical value greater or equal than 0. What is the maximum distance in fuzzy matching (i.e. Approximate String Distance, see adist()) between input place and match_by column to tolerate before testing alternative providers or looking for geographical matching with Nominatim API? This parameter is set equal to 0 if match_by is equal to iso3166_1_alpha2 or iso3166_2. Check Details and Examples in oe_match() to understand why this parameter is important. Ignored if place is not a character vector since the matching is performed through a spatial operation.

Value

A list with two elements, named url and file_size. The first element is the URL of the .osm.pbf file associated with the input place, while the second element is the size of the file in bytes (which may be NULL or NA)

Details

The fields iso3166_1_alpha2 and iso3166_2 are used by Geofabrik provider to perform matching operations using ISO 3166-1 alpha-2 and ISO 3166-2 codes. See geofabrik_zones for more details.

If the input place is specified as a spatial point (either as sfc_POINT or a pair of numeric coordinates), then the function will return the geographical area with the highest "level" intersecting the point. See the help pages of the chosen provider database to understand the meaning of the "level" field. If there are multiple areas at the same "level" intersecting the input place, then the function will return the area whose centroid is closer to the input place.

If the input place is specified as a character vector and there are multiple plausible matches between the input place and the match_by column, then the function will return a warning and it will select the first match. See Examples. On the other hand, if the approximate string distance between the input place and the best match in match_by column is greater than max_string_dist, then the function will look for exact matches (i.e. max_string_dist = 0) in the other supported providers. If it finds an exact match, then it will return the corresponding URL. Otherwise, if match_by is equal to "name", then it will try to geolocate the input place using the Nominatim API, and then it will perform a spatial matching operation (see examples and vignettes), while, if match_by != "name", then it will return an error.

See also

Examples

# The simplest example:
oe_match("Italy")

#> The input place was matched with: Italy

#> $url
#> [1] "https://download.geofabrik.de/europe/italy-latest.osm.pbf"
#> 
#> $file_size
#> [1] 1544340778
#> 

# The default provider is "geofabrik", but we can change that:
oe_match("Leeds", provider = "bbbike")

#> The input place was matched with: Leeds

#> $url
#> [1] "https://download.bbbike.org/osm/bbbike/Leeds/Leeds.osm.pbf"
#> 
#> $file_size
#> [1] 19376705
#> 

# By default, the matching operations are performed through the column "name"
# in the provider's database but this can be a problem. Hence, you can
# perform the matching operations using other columns:
oe_match("RU", match_by = "iso3166_1_alpha2")

#> The input place was matched with: RU

#> $url
#> [1] "https://download.geofabrik.de/russia-latest.osm.pbf"
#> 
#> $file_size
#> [1] 2820253009
#> 

# Run oe_providers() for reading a short description of all providers and
# check the help pages of the corresponding databases to learn which fields
# are present.

# You can always increase the max_string_dist argument, but it can be
# dangerous:
oe_match("London", max_string_dist = 3, quiet = FALSE)

#> The input place was matched with: Jordan

#> $url
#> [1] "https://download.geofabrik.de/asia/jordan-latest.osm.pbf"
#> 
#> $file_size
#> [1] 27400228
#> 

# Match the input zone using an sfc_POINT object:
milan_duomo = sf::st_sfc(sf::st_point(c(1514924, 5034552)), crs = 3003)
oe_match(milan_duomo, quiet = FALSE)

#> although coordinates are longitude/latitude, st_intersects assumes that they are planar

#> The input place was matched with multiple geographical areas. Selecting the areas with the highest "level". See the help page associated to the chosen provider for an explanation of the meaning of the "level" field.

#> $url
#> [1] "https://download.geofabrik.de/europe/italy/nord-ovest-latest.osm.pbf"
#> 
#> $file_size
#> [1] 416306623
#> 

leeds = sf::st_sfc(sf::st_point(c(430147.8, 433551.5)), crs = 27700)
oe_match(leeds, provider = "bbbike")

#> although coordinates are longitude/latitude, st_intersects assumes that they are planar

#> $url
#> [1] "https://download.bbbike.org/osm/bbbike/Leeds/Leeds.osm.pbf"
#> 
#> $file_size
#> [1] 19376705
#> 

# Match the input zone using a numeric vector of coordinates
# (in which case crs = 4326 is assumed)
oe_match(c(9.1916, 45.4650)) # Milan, Duomo using CRS = 4326

#> although coordinates are longitude/latitude, st_intersects assumes that they are planar

#> The input place was matched with multiple geographical areas. Selecting the areas with the highest "level". See the help page associated to the chosen provider for an explanation of the meaning of the "level" field.

#> $url
#> [1] "https://download.geofabrik.de/europe/italy/nord-ovest-latest.osm.pbf"
#> 
#> $file_size
#> [1] 416306623
#> 

# It returns a warning since Berin is matched both with Benin and Berlin
oe_match("Berin", quiet = FALSE)

#> Warning: The input place was matched with multiple geographical zones: Benin - Berlin. Selecting the first match.

#> The input place was matched with: Benin

#> $url
#> [1] "https://download.geofabrik.de/africa/benin-latest.osm.pbf"
#> 
#> $file_size
#> [1] 34426444
#> 

# If the input place does not match any zone in the chosen provider, then the
# function will test the other providers:
oe_match("Leeds")

#> No exact match found for place = Leeds and provider = geofabrik. Best match is Laos. 
#> Checking the other providers.

#> An exact string match was found using provider = bbbike.

#> $url
#> [1] "https://download.bbbike.org/osm/bbbike/Leeds/Leeds.osm.pbf"
#> 
#> $file_size
#> [1] 19376705
#> 

# If the input place cannot be exactly matched with any zone in any provider,
# then the function will try to geolocate the input and then it will perform a
# spatial match:
oe_match("Milan")

#> No exact match found for place = Milan and provider = geofabrik. Best match is Iran. 
#> Checking the other providers.

#> No exact match found in any OSM provider data. Searching for the location online.

#> although coordinates are longitude/latitude, st_intersects assumes that they are planar

#> The input place was matched with multiple geographical areas. Selecting the areas with the highest "level". See the help page associated to the chosen provider for an explanation of the meaning of the "level" field.

#> $url
#> [1] "https://download.geofabrik.de/europe/italy/nord-ovest-latest.osm.pbf"
#> 
#> $file_size
#> [1] 416306623
#>