Match input place with a URL

This function is used to match an input place with the URL of a .osm.pbf file (and its file-size, if present). The URLs are stored in several provider's databases. See oe_providers() and examples.

oe_match(place, ...)

# S3 method for default
oe_match(place, ...)

# S3 method for sfc_POINT
oe_match(place, provider = "geofabrik", quiet = TRUE, ...)

# S3 method for numeric
oe_match(place, provider = "geofabrik", quiet = TRUE, ...)

# S3 method for character
oe_match(
  place,
  provider = "geofabrik",
  match_by = "name",
  max_string_dist = 1,
  interactive_ask = FALSE,
  quiet = TRUE,
  ...
)

Arguments

place
Description of the geographical area that should be matched to a .osm.pbf file through the chosen provider. Can be either a length-1 character vector, a length-1 sfc_POINT object or a numeric vector of coordinates with length 2. In the latter case it is assumed that the EPSG code is 4326, while you can use any EPSG code with an sfc_POINT object. See details and examples of oe_match().

...
arguments passed to other methods

provider
Which provider should be used to download the data? Available providers can be found with the following command: oe_providers().

quiet
Boolean. If FALSE the function prints informative messages.

match_by
Which column of the provider's database should be used for matching the input place with a .osm.pbf file? The default is "name". Check details and examples to understand how this parameter works. Ignored if place is not a character vector since the matching is performed through a spatial operation.

max_string_dist
Numerical value greater or equal than 0. What is the maximum distance in fuzzy matching to tolerate before asking the user to select which zone to download? This parameter is set equal to 0 if match_by is equal to iso3166_1_alpha2 or iso3166_2. Check details and examples to understand why this parameter is important. Ignored if place is not a character vector since the matching is performed through a spatial operation.

interactive_ask
Boolean. If TRUE the function creates and interactive menu in case the best match is further than max_string_dist, otherwise it fails with stop(). Check details and examples to understand why this parameter is important. Ignored if place is not a character vector since the matching is performed through a spatial operation.

Value

A list with two elements, named url and file_size. The first element is the URL of the .osm.pbf file associated with the input place, while the second element is the size of the file in bytes (which may be NULL or NA)

Details

The fields iso3166_1_alpha2 and iso3166_2 are used by geofabrik provider to perform matching operations using ISO 3166-1 alpha-2 and ISO 3166-2. See ?geofabrik_zones for more details.

If the input place is specified as a spatial point (either sfc_POINT or numeric coordinates), then the function will return the geographical area with the highest "level" intersecting the point. See the help pages of the chosen provider database for understanding the meaning of the "level" field. If there are multiple areas at the same "level" intersecting the input place, then the function will return the area whose centroid is closer to the input place.

If the input place is specified as a character vector and there are multiple plausible matches between the input place and the match_by column, then the function will return a warning and it will select the first match. See examples.

See also

Examples

# The simplest example:
oe_match("Italy")

#> $url
#> [1] "https://download.geofabrik.de/europe/italy-latest.osm.pbf"
#> 
#> $file_size
#> [1] 1544340778
#> 

# The default provider is "geofabrik", but we can change that:
oe_match("Leeds", provider = "bbbike")

#> $url
#> [1] "https://download.bbbike.org/osm/bbbike/Leeds/Leeds.osm.pbf"
#> 
#> $file_size
#> [1] 19376705
#> 

# By default the matching operations are performed through the column "name"
# in the provider's database but this can be a problem:
if (FALSE) {
oe_match("Russia", quiet = FALSE)}
# So you can perform the matching operations using other columns in the
# provider's database:
oe_match("RU", match_by = "iso3166_1_alpha2")

#> $url
#> [1] "https://download.geofabrik.de/russia-latest.osm.pbf"
#> 
#> $file_size
#> [1] 2820253009
#> 

# Run oe_providers() for a description of all providers and check the help
# pages of the corresponding databases to learn which fields are present.

# You can always increase the max_string_dist argument to help the function:
if (FALSE) {
oe_match("Isle Wight", quiet = FALSE)}
oe_match("Isle Wight", max_string_dist = 3, quiet = FALSE)

#> The input place was matched with: Isle of Wight

#> $url
#> [1] "https://download.geofabrik.de/europe/great-britain/england/isle-of-wight-latest.osm.pbf"
#> 
#> $file_size
#> [1] 6877468
#> 

# but be aware that it can be dangerous:
oe_match("London", max_string_dist = 3, quiet = FALSE)

#> The input place was matched with: Jordan

#> $url
#> [1] "https://download.geofabrik.de/asia/jordan-latest.osm.pbf"
#> 
#> $file_size
#> [1] 27400228
#> 

# Match the input zone using an sfc_POINT object:
milan_duomo = sf::st_sfc(sf::st_point(c(1514924, 5034552)), crs = 3003)
oe_match(milan_duomo, quiet = FALSE)

#> although coordinates are longitude/latitude, st_intersects assumes that they are planar

#> The input place was matched with multiple geographical areas. Selecting the areas with the biggest "level". See the help page  associated to the chosen provider for an explanation of the meaning of the "level" field

#> $url
#> [1] "https://download.geofabrik.de/europe/italy/nord-ovest-latest.osm.pbf"
#> 
#> $file_size
#> [1] 416306623
#> 

leeds = sf::st_sfc(sf::st_point(c(430147.8, 433551.5)), crs = 27700)
oe_match(leeds, provider = "bbbike")

#> although coordinates are longitude/latitude, st_intersects assumes that they are planar

#> $url
#> [1] "https://download.bbbike.org/osm/bbbike/Leeds/Leeds.osm.pbf"
#> 
#> $file_size
#> [1] 19376705
#> 

# Match the input zone using a numeric vector of coordinates
# (in which case crs = 4326 is assumed)
oe_match(c(9.1916, 45.4650)) # Milan, Duomo using CRS = 4326

#> although coordinates are longitude/latitude, st_intersects assumes that they are planar

#> $url
#> [1] "https://download.geofabrik.de/europe/italy/nord-ovest-latest.osm.pbf"
#> 
#> $file_size
#> [1] 416306623
#> 

# Check interactive_ask:
if (interactive()) {
    oe_match("London", interactive_ask = TRUE)
}

# It returns a warning since Berin is matched both with Benin and Berlin
oe_match("Berin", quiet = FALSE)

#> Warning: The input place was matched with multiple geographical zones: Benin - Berlin. Selecting the first match.

#> The input place was matched with: Benin

#> $url
#> [1] "https://download.geofabrik.de/africa/benin-latest.osm.pbf"
#> 
#> $file_size
#> [1] 34426444
#>