Summary of available providers

This function is used to display a short summary of the major characteristics of the databases associated to all available providers.

oe_providers(quiet = FALSE)

Arguments

quiet

Boolean. If FALSE, the function prints informative messages. Starting from sf version 0.9.6, if quiet is equal to FALSE, then vectortranslate operations will display a progress bar.

Value

A data.frame with 4 columns representing the name of each available provider, the name of the corresponding database and the number of features and fields.

Examples

oe_providers()
#> Check the corresponding help pages to read more details about the fields in each database (e.g. ?geofabrik_zones).
#>   available_providers          database_name number_of_zones number_of_fields
#> 1           geofabrik        geofabrik_zones             476                8
#> 2              bbbike           bbbike_zones             237                5
#> 3    openstreetmap_fr openstreetmap_fr_zones            1187                6