Read a .pbf or .gpkg object from file or URL

This function is used to read a .pbf or .gpkg object from file or URL. It is a wrapper around oe_download(), oe_vectortranslate() and sf::st_read(), creating an easy way to download, convert, and read a .pbf or .gpkg file. Check the introductory vignette and the help pages of the wrapped function for more details.

oe_read(
  file_path,
  layer = "lines",
  ...,
  provider = NULL,
  download_directory = oe_download_directory(),
  file_size = NULL,
  force_download = FALSE,
  max_file_size = 5e+08,
  download_only = FALSE,
  skip_vectortranslate = FALSE,
  vectortranslate_options = NULL,
  osmconf_ini = NULL,
  extra_attributes = NULL,
  force_vectortranslate = NULL,
  quiet = TRUE
)

Arguments

file_path
A url or the path of a .osm.pbf or .gpkg file.

layer
Which layer should be read in? Typically points, lines (the default), multilinestrings, multipolygons or other_relations.

...
Arguments that should be passed to sf::st_read(), like query or stringsAsFactors.

provider
Which provider should be used to download the data? Available providers can be found with the following command: oe_providers().

download_directory
Where to download the file containing the OSM data? By default this is tempdir(), which changes each time you restart R. You can set a persistent download_directory by adding the following to your .Renviron file (e.g. with usethis::edit_r_environ()): OSMEXT_DOWNLOAD_DIRECTORY=/path/to/osm/data.

file_size
How big is the file? Optional. NA by default. If it's bigger than max_file_size and the function is run in interactive mode, then an interactive menu is displayed, asking for permission for downloading the file.

force_download
Should the file be updated if it has already been downloaded? FALSE by default. This parameter is used to update old .osm.pbf files.

max_file_size
The maximum file size to download without asking in interactive mode. Default: 5e+8, half a gigabyte.

download_only
Boolean. If TRUE then the function only returns the path where the matched file is stored, instead of reading it. FALSE by default.

skip_vectortranslate
Boolean. If TRUE then the function skips all the vectortranslate operations and it reads (or simply returns the path) of the .osm.pbf file. FALSE by default.

vectortranslate_options
Options to pass to the sf::gdal_utils() argument options. Set by default. Check Details at oe_vectortranslate().

osmconf_ini
The configuration file specifying which columns should be in the resulting data frame. See documentation at gdal.org. Check Details at oe_vectortranslate().

extra_attributes
Which addition columns, corresponding to OSM keys, should be in the resulting dataset? FALSE by default. Check Details at oe_vectortranslate() and oe_get_keys().

force_vectortranslate
Boolean. Force the original .pbf file to be translated into a .gpkg file, even if a .gpkg with the same name already exists? Check Details at oe_vectortranslate().

quiet
Boolean. If FALSE the function prints informative messages.

Value

An sf object.

Details

You cannot add any layer to an existing .gpkg file but you can extract some of the tags in other_tags. Check oe_get_keys() for more details.

The arguments provider, download_directory, file_size, force_download, and max_file_size are ignored if file_path points to an existing .pbf or .gpkg file.

Examples

# Read an existing .pbf file
my_pbf = system.file("its-example.osm.pbf", package = "osmextract")
oe_read(my_pbf)

#> Simple feature collection with 189 features and 9 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.562458 ymin: 53.80471 xmax: -1.548076 ymax: 53.81105
#> geographic CRS: WGS 84
#> First 10 features:
#>     osm_id           name     highway waterway aerialway barrier man_made
#> 1  4371081           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 2  4371084           <NA>    cycleway     <NA>      <NA>    <NA>     <NA>
#> 3  4419868    Cannon Walk     footway     <NA>      <NA>    <NA>     <NA>
#> 4  6277600 Cavendish Road     service     <NA>      <NA>    <NA>     <NA>
#> 5  6277601 Cavendish Road residential     <NA>      <NA>    <NA>     <NA>
#> 6  6295680  Blenheim Walk       trunk     <NA>      <NA>    <NA>     <NA>
#> 7  6962430           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 8  6962433           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 9  6962435           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 10 6962440  Cemetery Road     service     <NA>      <NA>    <NA>     <NA>
#>    z_order
#> 1        0
#> 2        0
#> 3        0
#> 4        0
#> 5        3
#> 6        8
#> 7        0
#> 8        0
#> 9        0
#> 10       0
#>                                                                                                                        other_tags
#> 1                                                                                                                            <NA>
#> 2                                                                                    "bicycle"=>"designated","foot"=>"designated"
#> 3                                                               "website"=>"http://woodhousemooronline.com/the-cannon-destroyer/"
#> 4                                                                             "access"=>"permissive","lanes"=>"1","oneway"=>"yes"
#> 5                                                                                                    "lanes"=>"1","oneway"=>"yes"
#> 6  "lanes"=>"2","lit"=>"yes","maxspeed"=>"30 mph","oneway"=>"yes","ref"=>"A660","turn:lanes"=>"left;through|through;slight_right"
#> 7                                                                                                                            <NA>
#> 8                                                                                                                            <NA>
#> 9                                                                                                                            <NA>
#> 10                                                                                                                           <NA>
#>                          geometry
#> 1  LINESTRING (-1.560083 53.80...
#> 2  LINESTRING (-1.559709 53.80...
#> 3  LINESTRING (-1.5609 53.8085...
#> 4  LINESTRING (-1.552587 53.80...
#> 5  LINESTRING (-1.551771 53.80...
#> 6  LINESTRING (-1.552894 53.80...
#> 7  LINESTRING (-1.554698 53.80...
#> 8  LINESTRING (-1.557451 53.80...
#> 9  LINESTRING (-1.556508 53.80...
#> 10 LINESTRING (-1.558149 53.80...

oe_read(my_pbf, layer = "points") # Read a new layer

#> Simple feature collection with 186 features and 10 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -1.568766 ymin: 53.80569 xmax: -1.549451 ymax: 53.81136
#> geographic CRS: WGS 84
#> First 10 features:
#>      osm_id                         name barrier         highway      ref
#> 1  21069418                         <NA>    <NA> traffic_signals     <NA>
#> 2  21093230                         <NA>    <NA> mini_roundabout     <NA>
#> 3  26653419 Statue of Duke of Wellington    <NA>            <NA>     <NA>
#> 4  31004252                         <NA>    <NA> traffic_signals     <NA>
#> 5  31004259                         <NA>    <NA> traffic_signals     <NA>
#> 6  31004270                         <NA>    <NA>        crossing     <NA>
#> 7  31004287                         <NA>    <NA> traffic_signals     <NA>
#> 8  52905112                         <NA>    <NA>            <NA>   LS2 54
#> 9  52905119                         <NA>    <NA>            <NA>     <NA>
#> 10 52905122       Leeds University Steps    <NA>        bus_stop 45011386
#>    address is_in place man_made
#> 1     <NA>  <NA>  <NA>     <NA>
#> 2     <NA>  <NA>  <NA>     <NA>
#> 3     <NA>  <NA>  <NA>     <NA>
#> 4     <NA>  <NA>  <NA>     <NA>
#> 5     <NA>  <NA>  <NA>     <NA>
#> 6     <NA>  <NA>  <NA>     <NA>
#> 7     <NA>  <NA>  <NA>     <NA>
#> 8     <NA>  <NA>  <NA>     <NA>
#> 9     <NA>  <NA>  <NA>     <NA>
#> 10    <NA>  <NA>  <NA>     <NA>
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        other_tags
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                         "crossing"=>"traffic_signals","crossing_ref"=>"pelican"
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            <NA>
#> 3                                                                                                                                                                                                                                                                                                                                                                                       "historic"=>"memorial","tourism"=>"artwork","url"=>"http://www.leodis.net/display.aspx?resourceIdentifier=2008118_165852"
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                  "crossing"=>"traffic_signals","crossing_ref"=>"toucan","kerb"=>"flush","tactile_paving"=>"yes"
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            <NA>
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           "crossing"=>"pelican"
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                         "crossing"=>"traffic_signals","crossing_ref"=>"pelican"
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                                         "amenity"=>"post_box","operator"=>"Royal Mail","post_box:type"=>"meter"
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                    "amenity"=>"telephone","booth"=>"KXPlus","covered"=>"booth","operator"=>"BT"
#> 10 "bus"=>"yes","local_ref"=>"45011386","naptan:AtcoCode"=>"450011386","naptan:Bearing"=>"N","naptan:CommonName"=>"Leeds University","naptan:Crossing"=>"Back Blenheim Terrace","naptan:Indicator"=>"Stop 45011386","naptan:Landmark"=>"Leeds University","naptan:Notes"=>"New TF shelter and 3P fitted awaiting survey","naptan:PlusbusZoneRef"=>"LEEDS","naptan:ShortCommonName"=>"Leeds University","naptan:Street"=>"Woodhouse Lane","naptan:verified"=>"yes","public_transport"=>"platform","shelter"=>"yes"
#>                      geometry
#> 1   POINT (-1.552508 53.8086)
#> 2   POINT (-1.56583 53.81002)
#> 3  POINT (-1.559813 53.80819)
#> 4   POINT (-1.56048 53.80638)
#> 5  POINT (-1.559558 53.80796)
#> 6  POINT (-1.559285 53.80795)
#> 7  POINT (-1.551964 53.80751)
#> 8   POINT (-1.55062 53.80642)
#> 9  POINT (-1.552372 53.80799)
#> 10 POINT (-1.552314 53.80809)

oe_read(my_pbf, extra_attributes = c("oneway", "ref")) # Add new tags

#> Simple feature collection with 189 features and 11 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.562458 ymin: 53.80471 xmax: -1.548076 ymax: 53.81105
#> geographic CRS: WGS 84
#> First 10 features:
#>     osm_id           name     highway waterway aerialway barrier man_made
#> 1  4371081           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 2  4371084           <NA>    cycleway     <NA>      <NA>    <NA>     <NA>
#> 3  4419868    Cannon Walk     footway     <NA>      <NA>    <NA>     <NA>
#> 4  6277600 Cavendish Road     service     <NA>      <NA>    <NA>     <NA>
#> 5  6277601 Cavendish Road residential     <NA>      <NA>    <NA>     <NA>
#> 6  6295680  Blenheim Walk       trunk     <NA>      <NA>    <NA>     <NA>
#> 7  6962430           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 8  6962433           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 9  6962435           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 10 6962440  Cemetery Road     service     <NA>      <NA>    <NA>     <NA>
#>    oneway  ref z_order
#> 1    <NA> <NA>       0
#> 2    <NA> <NA>       0
#> 3    <NA> <NA>       0
#> 4     yes <NA>       0
#> 5     yes <NA>       3
#> 6     yes A660       8
#> 7    <NA> <NA>       0
#> 8    <NA> <NA>       0
#> 9    <NA> <NA>       0
#> 10   <NA> <NA>       0
#>                                                                                          other_tags
#> 1                                                                                              <NA>
#> 2                                                      "bicycle"=>"designated","foot"=>"designated"
#> 3                                 "website"=>"http://woodhousemooronline.com/the-cannon-destroyer/"
#> 4                                                               "access"=>"permissive","lanes"=>"1"
#> 5                                                                                      "lanes"=>"1"
#> 6  "lanes"=>"2","lit"=>"yes","maxspeed"=>"30 mph","turn:lanes"=>"left;through|through;slight_right"
#> 7                                                                                              <NA>
#> 8                                                                                              <NA>
#> 9                                                                                              <NA>
#> 10                                                                                             <NA>
#>                          geometry
#> 1  LINESTRING (-1.560083 53.80...
#> 2  LINESTRING (-1.559709 53.80...
#> 3  LINESTRING (-1.5609 53.8085...
#> 4  LINESTRING (-1.552587 53.80...
#> 5  LINESTRING (-1.551771 53.80...
#> 6  LINESTRING (-1.552894 53.80...
#> 7  LINESTRING (-1.554698 53.80...
#> 8  LINESTRING (-1.557451 53.80...
#> 9  LINESTRING (-1.556508 53.80...
#> 10 LINESTRING (-1.558149 53.80...

# Read an existing .gpkg file. This file was created by oe_read
my_gpkg = system.file("its-example.gpkg", package = "osmextract")
oe_read(my_gpkg)

#> Simple feature collection with 189 features and 11 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.562458 ymin: 53.80471 xmax: -1.548076 ymax: 53.81105
#> geographic CRS: WGS 84
#> First 10 features:
#>     osm_id           name     highway waterway aerialway barrier man_made
#> 1  4371081           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 2  4371084           <NA>    cycleway     <NA>      <NA>    <NA>     <NA>
#> 3  4419868    Cannon Walk     footway     <NA>      <NA>    <NA>     <NA>
#> 4  6277600 Cavendish Road     service     <NA>      <NA>    <NA>     <NA>
#> 5  6277601 Cavendish Road residential     <NA>      <NA>    <NA>     <NA>
#> 6  6295680  Blenheim Walk       trunk     <NA>      <NA>    <NA>     <NA>
#> 7  6962430           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 8  6962433           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 9  6962435           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 10 6962440  Cemetery Road     service     <NA>      <NA>    <NA>     <NA>
#>    oneway  ref z_order
#> 1    <NA> <NA>       0
#> 2    <NA> <NA>       0
#> 3    <NA> <NA>       0
#> 4     yes <NA>       0
#> 5     yes <NA>       3
#> 6     yes A660       8
#> 7    <NA> <NA>       0
#> 8    <NA> <NA>       0
#> 9    <NA> <NA>       0
#> 10   <NA> <NA>       0
#>                                                                                          other_tags
#> 1                                                                                              <NA>
#> 2                                                      "bicycle"=>"designated","foot"=>"designated"
#> 3                                 "website"=>"http://woodhousemooronline.com/the-cannon-destroyer/"
#> 4                                                               "access"=>"permissive","lanes"=>"1"
#> 5                                                                                      "lanes"=>"1"
#> 6  "lanes"=>"2","lit"=>"yes","maxspeed"=>"30 mph","turn:lanes"=>"left;through|through;slight_right"
#> 7                                                                                              <NA>
#> 8                                                                                              <NA>
#> 9                                                                                              <NA>
#> 10                                                                                             <NA>
#>                          geometry
#> 1  LINESTRING (-1.560083 53.80...
#> 2  LINESTRING (-1.559709 53.80...
#> 3  LINESTRING (-1.5609 53.8085...
#> 4  LINESTRING (-1.552587 53.80...
#> 5  LINESTRING (-1.551771 53.80...
#> 6  LINESTRING (-1.552894 53.80...
#> 7  LINESTRING (-1.554698 53.80...
#> 8  LINESTRING (-1.557451 53.80...
#> 9  LINESTRING (-1.556508 53.80...
#> 10 LINESTRING (-1.558149 53.80...

# You cannot add any layer to an existing .gpkg file but you can extract some
# of the tags in other_tags. Check oe_get_keys() for more details.
names(oe_read(my_gpkg, extra_attributes = c("maxspeed")))

#>  [1] "osm_id"     "name"       "highway"    "waterway"   "aerialway" 
#>  [6] "barrier"    "man_made"   "oneway"     "ref"        "z_order"   
#> [11] "other_tags" "geometry"  

# Delete the .gpkg file not to mess with other examples
file.remove(my_gpkg)

#> [1] TRUE

# Read from a URL
my_url = "https://github.com/ITSLeeds/osmextract/raw/master/inst/its-example.osm.pbf"
# Please note that if you read from a URL which is not linked to one of the
# supported providers, you need to specify the provider parameter:
if (FALSE) {
oe_read(my_url, provider = "test")
}