Read a .pbf or .gpkg object

Read a .pbf or .gpkg object

oe_read(
  file_path,
  layer = "lines",
  ...,
  download_directory = oe_download_directory(),
  vectortranslate_options = NULL,
  osmconf_ini = NULL,
  extra_attributes = NULL,
  force_vectortranslate = NULL,
  skip_vectortranslate = FALSE,
  quiet = TRUE
)

Arguments

file_path
The path of the .gpkg or .osm.pbf file that should be read-in.

layer
Which layer should be read in? Typically points, lines (the default), multilinestrings, multipolygons or other_relations.

...
Arguments that should be passed to sf::st_read()

download_directory
Where to download the file containing the OSM data? By default this is tempdir(), which changes each time you restart R. You can set a persistent download_directory by adding the following to your .Renviron file (e.g. with usethis::edit_r_environ()): OSMEXT_DOWNLOAD_DIRECTORY=/path/to/osm/data.

vectortranslate_options
Options to pass to the sf::gdal_utils() argument options. Set by default. Check details and examples.

osmconf_ini
The configuration file specifying which columns should be in the resulting data frame. See documentation at gdal.org for details.

extra_attributes
Which addition columns, corresponding to OSM keys, should be in the resulting dataset? NULL by default.

force_vectortranslate
Force the original .pbf file to be translated into a .gpkg file, even if a .gpkg associated with the provider zone already exists.

skip_vectortranslate
Boolean. If TRUE then the function skips all the vectortranslate operations and it reads (or simply returns the path) of the .osm.pbf file. FALSE by default.

quiet
Boolean. If FALSE the function prints informative messages.

Value

An sf object related to the input path.

Details

For the moment I will consider only .gpkg or .osm.pbf file. The approach is more or less the same as for oe_get() but we need to skip the matching operations. In the near future I think I could merge the two approaches i.e. oe_get will be oe_match + oe_read to avoid a lot of duplication.

Examples

oe_read(
  file_path = system.file("its-example.osm.pbf", package = "osmextract"),
  download_directory = tempdir()
)

#> Simple feature collection with 189 features and 9 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.562458 ymin: 53.80471 xmax: -1.548076 ymax: 53.81105
#> geographic CRS: WGS 84
#> First 10 features:
#>     osm_id           name     highway waterway aerialway barrier man_made
#> 1  4371081           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 2  4371084           <NA>    cycleway     <NA>      <NA>    <NA>     <NA>
#> 3  4419868    Cannon Walk     footway     <NA>      <NA>    <NA>     <NA>
#> 4  6277600 Cavendish Road     service     <NA>      <NA>    <NA>     <NA>
#> 5  6277601 Cavendish Road residential     <NA>      <NA>    <NA>     <NA>
#> 6  6295680  Blenheim Walk       trunk     <NA>      <NA>    <NA>     <NA>
#> 7  6962430           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 8  6962433           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 9  6962435           <NA>     footway     <NA>      <NA>    <NA>     <NA>
#> 10 6962440  Cemetery Road     service     <NA>      <NA>    <NA>     <NA>
#>    z_order
#> 1        0
#> 2        0
#> 3        0
#> 4        0
#> 5        3
#> 6        8
#> 7        0
#> 8        0
#> 9        0
#> 10       0
#>                                                                                                                        other_tags
#> 1                                                                                                                            <NA>
#> 2                                                                                    "bicycle"=>"designated","foot"=>"designated"
#> 3                                                               "website"=>"http://woodhousemooronline.com/the-cannon-destroyer/"
#> 4                                                                             "access"=>"permissive","lanes"=>"1","oneway"=>"yes"
#> 5                                                                                                    "lanes"=>"1","oneway"=>"yes"
#> 6  "lanes"=>"2","lit"=>"yes","maxspeed"=>"30 mph","oneway"=>"yes","ref"=>"A660","turn:lanes"=>"left;through|through;slight_right"
#> 7                                                                                                                            <NA>
#> 8                                                                                                                            <NA>
#> 9                                                                                                                            <NA>
#> 10                                                                                                                           <NA>
#>                          geometry
#> 1  LINESTRING (-1.560083 53.80...
#> 2  LINESTRING (-1.559709 53.80...
#> 3  LINESTRING (-1.5609 53.8085...
#> 4  LINESTRING (-1.552587 53.80...
#> 5  LINESTRING (-1.551771 53.80...
#> 6  LINESTRING (-1.552894 53.80...
#> 7  LINESTRING (-1.554698 53.80...
#> 8  LINESTRING (-1.557451 53.80...
#> 9  LINESTRING (-1.556508 53.80...
#> 10 LINESTRING (-1.558149 53.80...