Translate the .osm.pbf format into .gpkg

Translate the .osm.pbf format into .gpkg

oe_vectortranslate(
  file_path,
  vectortranslate_options = NULL,
  layer = NULL,
  osmconf_ini = NULL,
  extra_attributes = NULL,
  force_vectortranslate = FALSE,
  quiet = TRUE
)

Arguments

file_path
Character string representing the full path the .osm.pbf file

vectortranslate_options
Options to pass to the sf::gdal_utils() argument options. Set by default. Check details and examples.

layer
Which layer should be read in? Typically points, lines (the default), multilinestrings, multipolygons or other_relations.

osmconf_ini
The configuration file specifying which columns should be in the resulting data frame. See documentation at gdal.org for details.

extra_attributes
Which addition columns, corresponding to OSM keys, should be in the resulting dataset? NULL by default.

force_vectortranslate
Force the original .pbf file to be translated into a .gpkg file, even if a .gpkg associated with the provider zone already exists.

quiet
Boolean. If FALSE the function prints informative messages.

Value

Character string representing the full path the .gpkg file

Examples

# First we need to match an input zone with a .osm.pbf file
its_match = oe_match("ITS Leeds", provider = "test")
# The we can download the .osm.pbf files
its_pbf = oe_download(
  file_url = its_match$url,
  file_size = its_match$file_size,
  download_directory = tempdir(),
  provider = "test"
)
list.files(tempdir(), pattern = "pbf|gpkg")

#> [1] "test_its-example.gpkg"    "test_its-example.osm.pbf"

its_gpkg = oe_vectortranslate(
 its_pbf
)
list.files(tempdir(), pattern = "pbf|gpkg")

#> [1] "test_its-example.gpkg"    "test_its-example.osm.pbf"